Crate jailer

Modules

cgroup ğŸ”’

chroot ğŸ”’

env ğŸ”’

resource_limits ğŸ”’

Enums

JailerError

Constants

JAILER_VERSION ğŸ”’

Functions

build_arg_parser

Create an ArgParser object which contains info about the command line argument parser and populate it with the expected arguments and their characteristics.

clean_env_vars ğŸ”’

close_fds_by_close_range ğŸ”’

close_fds_by_reading_proc ğŸ”’

close_inherited_fds ğŸ”’

main ğŸ”’

main_exec ğŸ”’

readln_special

sanitize_process ğŸ”’

to_cstring

Turns an AsRef<Path> into a CString (c style string). The expect should not fail, since Linux paths only contain valid Unicode chars (do they?), and do not contain null bytes (do they?).

writeln_special