Crate jailer

Crate for creating sandbox environment and do some action within a sandbox environment

Structs

EnvJailer

EnvJailer struct which creates jail. EnvJailer is build on top of Jailer which also handles environment variable. It is different than Jailer since environment variable set and unset operation is unsafe

Jailer

Jailer struct which creates jail.