jailer

Struct EnvJailer

pub struct EnvJailer { /* private fields */ }

EnvJailer struct which creates jail. EnvJailer is build on top of Jailer which also handles environment variable. It is different than Jailer since environment variable set and unset operation is unsafe

EnvJailer reverts to original env variable when it is dropped or closed

Implementations

impl EnvJailer

pub fn new() -> Result<Self, Error>

Create new EnvJailer

Errors

if new EnvJailer cannot be created

Example

use jailer::EnvJailer;

let mut env_jailer = EnvJailer::new().unwrap();
// do some action in jailer
unsafe {
    env_jailer.close().unwrap();
}

pub fn set_preserved_env<K>(&mut self, key: K)
where K: AsRef<OsStr>,

Set provided environment variable name as preserved environment variable. Saving environment variable content when Jailer gets dropped

Example

use jailer::EnvJailer;
let mut env_jailer = EnvJailer::new().unwrap();
unsafe {
    std::env::set_var("KEY", "VALUE");
    std::env::set_var("KEY2", "VALUE2");
}
env_jailer.set_preserved_env("KEY");
assert_eq!(std::env::var("KEY"), Ok("VALUE".to_string()));
unsafe {
    std::env::set_var("KEY", "VALUE2");
}
assert_eq!(std::env::var("KEY"), Ok("VALUE2".to_string()));
unsafe {
    env_jailer.close().unwrap();
}
assert_eq!(std::env::var("KEY"), Ok("VALUE2".to_string()));
assert!(std::env::var("KEY2").is_err());

pub fn remove_preserved_env<K>(&mut self, key: K)
where K: AsRef<OsStr>,

Remove environment variable from preserved env

This function do not remove current environment variable. To remove current environment variable you need to manually call std::env::remove_var.

Example

use jailer::EnvJailer;

let mut env_jailer = EnvJailer::new().unwrap();
unsafe {
    std::env::set_var("KEY", "VALUE");
}
env_jailer.set_preserved_env("KEY");
assert_eq!(std::env::var("KEY"), Ok("VALUE".to_string()));
env_jailer.remove_preserved_env("KEY");
unsafe {
    env_jailer.close().unwrap();
}
assert!(std::env::var("KEY").is_err());

pub fn original_env_vars_os(&self) -> &HashMap<OsString, OsString>

Return hashmap of original env variables

If any env is added by using set env than those env are not provided in this response use EnvJailer::preserved_env_vars_os

pub fn preserved_env_vars_os(&self) -> &HashSet<OsString>

Return hash set of preserved env variables

pub unsafe fn close(&mut self) -> Result<(), Error>

Closes a EnvJailer

Although EnvJailer removes directory and environment variable on drop, It may not remove directory or change current directory which can still fails and error will be ignored. To detect and handle error due to change of directory or deletion of temp dir call close manually

While closing/dropping, At first all current environment variables are removed than original env variables gets added at last preserved env variables gets added in those order before changing to original directory and closing temporary directory at last

Errors

When EnvJailer cannot be closed properly

Safety

Close function calls std::env::remove_var and std::env::set_var function which are both unsafe operation